Breaks converters up into seperate module / directory structure and supplants imgkit with grover
