+++ lib/inkblot/converters.rb
@@ -1,0 +1,8 @@
+module Inkblot
+	require 'inkblot/converters/converter'
+	require 'inkblot/converters/image_converter'
+	require 'inkblot/converters/html_converter'
+
+	# Grouped classes for converting content
+	module Converters; end
+end
